LibraryCallMonitor: don't use module name to track library calls
LibraryCallMonitor uses ProcessExecutionDetector and ModuleMap to track library
calls made in a particular process. However, there is an inconsistency between
the module name that ModuleMap uses and the process name that
ProcessExecutionDetector uses, which makes it impossible to track library calls
this way. So let's not use the module name!