Makefile: refactor
Removed stamp llvm-native-make; this target doesn't really make sense given
that we download the LLVM binary. Renamed to clang-binary and reshuffled the
Makefile around this.

Removed the clean and docs targets - they are waaaay out of date and no longer
work.

Also renamed some of the variables by adding underscores (which makes them
easier to read).