malware-inject: Wait for all processes spawned by the malware process
For redaman, we've been running into an issue where the malware process spawned by malware-inject terminates without waiting for child processes it has spawned. This causes malware-inject to exit, which then leads to the bootstrap.sh script exiting, and s2e takes that as a signal to kill the current state.

This commit makes malware-inject wait for the entire "process tree" rooted at the first malware process.